+++ src/neoxp/Node/NodeUtility.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Numerics;
 using System.Threading.Tasks;
 using Neo;
 using Neo.BlockchainToolkit.Models;
@@ -9,17 +10,19 @@ using Neo.Cryptography;
 using Neo.Cryptography.ECC;
 using Neo.IO;
 using Neo.Network.P2P.Payloads;
+using Neo.Network.RPC;
 using Neo.Persistence;
 using Neo.SmartContract;
 using Neo.SmartContract.Manifest;
 using Neo.SmartContract.Native;
 using Neo.VM;
 using Neo.Wallets;
+using NeoExpress.Commands;
 using NeoExpress.Models;
 
 namespace NeoExpress.Node
 {
-    class ExpressOracle
+    class NodeUtility
     {
         public static Block CreateSignedBlock(Header prevHeader, IReadOnlyList<KeyPair> keyPairs, uint network, Transaction[]? transactions = null, ulong timestamp = 0)
         {
@@ -69,7 +72,7 @@ namespace NeoExpress.Node
 
             if (blockCount == 1)
             {
-                var block = ExpressOracle.CreateSignedBlock(
+                var block = CreateSignedBlock(
                     prevHeader, keyPairs, network, timestamp: timestamp + delta);
                 await submitBlockAsync(block).ConfigureAwait(false);
             }
@@ -78,7 +81,7 @@ namespace NeoExpress.Node
                 var period = delta / (blockCount - 1);
                 for (int i = 0; i < blockCount; i++)
                 {
-                    var block = ExpressOracle.CreateSignedBlock(
+                    var block = CreateSignedBlock(
                         prevHeader, keyPairs, network, timestamp: timestamp);
                     await submitBlockAsync(block).ConfigureAwait(false);
                     prevHeader = block.Header;
@@ -208,10 +211,207 @@ namespace NeoExpress.Node
             return tx;
         }
 
+        // constants from ContractManagement native contracts
+        const byte Prefix_Contract = 8;
+        const byte Prefix_NextAvailableId = 15;
+
+        public static async Task<(ContractState contractState, IReadOnlyList<(string key, string value)> storagePairs)> DownloadContractStateAsync(
+                string contractHash, string rpcUri, uint stateHeight)
+        {
+            if (!UInt160.TryParse(contractHash, out var _contractHash))
+            {
+                throw new ArgumentException($"Invalid contract hash: \"{contractHash}\"");
+            }
+
+            if (!TransactionExecutor.TryParseRpcUri(rpcUri, out var uri))
+            {
+                throw new ArgumentException($"Invalid RpcUri value \"{rpcUri}\"");
+            }
+
+            using var rpcClient = new RpcClient(uri);
+            var stateAPI = new StateAPI(rpcClient);
+
+            if (stateHeight == 0)
+            {
+                uint? localRootIndex;
+                try
+                {
+                    (localRootIndex, _) = await stateAPI.GetStateHeightAsync();
+                }
+                catch (RpcException e)
+                {
+                    if (e.Message.Contains("Method not found"))
+                    {
+                        throw new Exception(
+                            "Could not get state information. Make sure the remote RPC server has state service support");
+                    }
+                    throw;
+                }
+
+                stateHeight = localRootIndex.HasValue ? localRootIndex.Value
+                    : throw new Exception($"Null \"{nameof(localRootIndex)}\" in state height response");
+            }
+
+            var stateRoot = await stateAPI.GetStateRootAsync(stateHeight);
+
+            // rpcClient.GetContractStateAsync returns the current ContractState, but this method needs
+            // the ContractState as it was at stateHeight. ContractManagement stores ContractState by
+            // contractHash with the prefix 8. The following code uses stateAPI.GetStateAsync to retrieve
+            // the value with that key at the height state root and then deserializes it into a ContractState
+            // instance via GetInteroperable.
+
+            var key = new byte[21];
+            key[0] = Prefix_Contract;
+            _contractHash.ToArray().CopyTo(key, 1);
+
+            ContractState contractState;
+            try
+            {
+                var contractStateBuffer = await stateAPI.GetStateAsync(
+                    stateRoot.RootHash, NativeContract.ContractManagement.Hash, key);
+                contractState = new StorageItem(contractStateBuffer).GetInteroperable<ContractState>();
+            }
+            catch (RpcException ex)
+            {
+                const int COR_E_KEYNOTFOUND = unchecked((int)0x80131577);
+                if (ex.HResult == COR_E_KEYNOTFOUND) throw new Exception($"Contract {contractHash} not found at height {stateHeight}");
+                throw;
+            }
+
+            if (contractState.Id < 0) throw new NotSupportedException("Contract download not supported for native contracts");
+
+            var states = await rpcClient.ExpressFindStatesAsync(stateRoot.RootHash, _contractHash, default);
+
+            return (contractState, states);
+        }
+
+        public static int PersistContract(NeoSystem neoSystem, ContractState state,
+            IReadOnlyList<(string key, string value)> storagePairs, ContractCommand.OverwriteForce force)
+        {
+            if (state.Id < 0) throw new ArgumentException("PersistContract not supported for native contracts", nameof(state));
+
+            using var snapshot = neoSystem.GetSnapshot();
+
+            StorageKey key = new KeyBuilder(NativeContract.ContractManagement.Id, Prefix_Contract).Add(state.Hash);
+            var localContract = snapshot.GetAndChange(key)?.GetInteroperable<ContractState>();
+            if (localContract is null)
+            {
+                // if localContract is null, the downloaded contract does not exist in the local Express chain
+                // Save the downloaded state + storage directly to the local chain
+
+                state.Id = GetNextAvailableId(snapshot);
+                snapshot.Add(key, new StorageItem(state));
+                PersistStoragePairs(snapshot, state.Id, storagePairs);
+
+                snapshot.Commit();
+                return state.Id;
+            }
+
+            // if localContract is not null, compare the current state + storage to the downloaded state + storage
+            // and overwrite changes if specified by user option
+
+            var (overwriteContract, overwriteStorage) = force switch
+            {
+                ContractCommand.OverwriteForce.All => (true, true),
+                ContractCommand.OverwriteForce.ContractOnly => (true, false),
+                ContractCommand.OverwriteForce.None => (false, false),
+                ContractCommand.OverwriteForce.StorageOnly => (false, true),
+                _ => throw new NotSupportedException($"Invalid OverwriteForce value {force}"),
+            };
+
+            var dirty = false;
+
+            if (!ContractStateEquals(state, localContract))
+            {
+                if (overwriteContract)
+                {
+                    // Note: a ManagementContract.Update() will not change the contract hash. Not even if the NEF changed.
+                    localContract.Nef = state.Nef;
+                    localContract.Manifest = state.Manifest;
+                    localContract.UpdateCounter = state.UpdateCounter;
+                    dirty = true;
+                }
+                else
+                {
+                    throw new Exception("Downloaded contract already exists. Use --force to overwrite");
+                }
+            }
+
+            if (!ContractStorageEquals(localContract.Id, snapshot, storagePairs))
+            {
+                if (overwriteStorage)
+                {
+                    byte[] prefix_key = StorageKey.CreateSearchPrefix(localContract.Id, default);
+                    foreach (var (k, v) in snapshot.Find(prefix_key))
+                    {
+                        snapshot.Delete(k);
+                    }
+                    PersistStoragePairs(snapshot, localContract.Id, storagePairs);
+                    dirty = true;
+                }
+                else
+                {
+                    throw new Exception("Downloaded contract storage already exists. Use --force to overwrite");
+                }
+            }
+
+            if (dirty) snapshot.Commit();
+            return localContract.Id;
+
+            static int GetNextAvailableId(DataCache snapshot)
+            {
+                StorageKey key = new KeyBuilder(NativeContract.ContractManagement.Id, Prefix_NextAvailableId);
+                StorageItem item = snapshot.GetAndChange(key);
+                int value = (int)(BigInteger)item;
+                item.Add(1);
+                return value;
+            }
+
+            static void PersistStoragePairs(DataCache snapshot, int contractId, IReadOnlyList<(string key, string value)> storagePairs)
+            {
+                for (int i = 0; i < storagePairs.Count; i++)
+                {
+                    snapshot.Add(
+                        new StorageKey { Id = contractId, Key = Convert.FromBase64String(storagePairs[i].key) },
+                        new StorageItem(Convert.FromBase64String(storagePairs[i].value)));
+                }
+            }
+
+            static bool ContractStateEquals(ContractState a, ContractState b)
+            {
+                return a.Hash.Equals(b.Hash)
+                    && a.UpdateCounter == b.UpdateCounter
+                    && a.Nef.ToArray().SequenceEqual(b.Nef.ToArray())
+                    && a.Manifest.ToJson().ToByteArray(false).SequenceEqual(b.Manifest.ToJson().ToByteArray(false));
+            }
+
+            static bool ContractStorageEquals(int contractId, DataCache snapshot, IReadOnlyList<(string key, string value)> storagePairs)
+            {
+                IReadOnlyDictionary<string, string> storagePairMap = storagePairs.ToDictionary(p => p.key, p => p.value);
+                var storageCount = 0;
+
+                byte[] prefixKey = StorageKey.CreateSearchPrefix(contractId, default);
+                foreach (var (k, v) in snapshot.Find(prefixKey))
+                {
+                    var storageKey = Convert.ToBase64String(k.Key);
+                    if (storagePairMap.TryGetValue(storageKey, out var storageValue)
+                        && storageValue.Equals(Convert.ToBase64String(v.Value)))
+                    {
+                        storageCount++;
+                    }
+                    else
+                    {
+                        return false;
+                    }
+                }
+
+                return storageCount != storagePairs.Count;
+            }
+        }
+
         // Need an IVerifiable.GetScriptHashesForVerifying implementation that doesn't
         // depend on the DataCache snapshot parameter in order to create a 
         // ContractParametersContext without direct access to node data.
-
         class BlockScriptHashes : IVerifiable
         {
             readonly UInt160[] hashes;